Add support for arbitrary attribute for img
This is direct support for article as appeared on wechat platform.
Image has data-src attribute which the original code of html-to-text
doesn't support.

As it's diverse from the original intension and standard, thus this is
meant to be changed or updated later as appropriated.

If your project needs this support then change your package.json to pull
down html-to-text via
"git+ssh://git@github.com/haxpor/node-html-to-text.git#fkit_compat".